requirements.txt
Flask==3.0.3
flask-cors==4.0.0
google-genai>=0.2.0
gunicorn==22.0.0
python-dotenv==1.0.1
